port placeholder_deploy test
